#1505 reformatted model but getting different answers
