+++ docs/source/models/submodels/porosity/index.rst
@@ -7,4 +7,5 @@ Porosity
   base_porosity
   constant_porosity
   reaction_driven_porosity
+  reaction_driven_porosity_ode
