+++ docs/source/models/submodels/porosity/reaction_driven_porosity_ode.rst
@@ -1,0 +1,8 @@
+Reaction-driven Model as an ODE
+===============================
+
+.. autoclass:: pybamm.porosity.ReactionDrivenODE
+    :members:
+
+
+
