+++ pybamm/models/full_battery_models/lithium_ion/base_lithium_ion_model.py
@@ -244,6 +244,13 @@ class BaseModel(pybamm.BaseBatteryModel):
             self.options["SEI porosity change"] == "true"
             or self.options["lithium plating porosity change"] == "true"
         ):
+            self.submodels["porosity"] = pybamm.porosity.ReactionDrivenODE(
+                self.param, self.options, self.x_average
+            )
+        elif (
+            self.options["SEI porosity change"] == "true2"
+            or self.options["lithium plating porosity change"] == "true"
+        ):
             self.submodels["porosity"] = pybamm.porosity.ReactionDriven(
                 self.param, self.options, self.x_average
             )
