+++ pybamm/models/submodels/porosity/__init__.py
@@ -1,3 +1,4 @@
 from .base_porosity import BaseModel
 from .constant_porosity import Constant
 from .reaction_driven_porosity import ReactionDriven
+from .reaction_driven_porosity_ode import ReactionDrivenODE
